+++ changelog.md
@@ -5,7 +5,7 @@
 ### 1.5.2, soon...
 
 * fixed: T2T Toolkit breaks posted Gravity Forms total field when products have options
-* changed: when Use Sandbox setting is selected, PxPay endpoint for UAT is used instead of SEC endpoint
+* changed: when Use Sandbox setting is selected, PxPay endpoint for UAT can be used instead of SEC endpoint
 
 ### 1.5.1, 2015-12-27